自定义跑步距离过短的AlertView，写了一点跑步首页的功能

+++ MRMobileRun/MRRunning/RunMainPageCV.h
@@ -23,7 +23,7 @@ typedef enum : NSUInteger {
 @property(nonatomic,assign)SportsState sportsState;
 
  // 当前跑步距离
-@property(nonatomic,assign)NSInteger distance;
+@property(nonatomic,assign)double distance;
